usb: renesas_usbhs: Use dmac only if the pipe type is bulk
This patch fixes an issue that isochronous transfer's data is possible to
be lost as a workaround. Since this driver uses a workqueue to start
the dmac, the transfer is possible to be delayed when system load is high.

Fixes: 6e4b74e4690d ("usb: renesas: fix scheduling in atomic context bug")
Cc: <stable@vger.kernel.org> # v3.4+
Signed-off-by: Yoshihiro Shimoda <yoshihiro.shimoda.uh@renesas.com>
Signed-off-by: Felipe Balbi <felipe.balbi@linux.intel.com>
(cherry picked from commit 700aa7ff8d2c2b9cc669c99375e2ccd06d3cd38d)
Signed-off-by: Kazuya Mizuguchi <kazuya.mizuguchi.ks@renesas.com>