mmc: tmio: fix soft lockup on CMD12 for R-Car SDHI
For regular multi block transfers the driver will use the automatic
CMD12 of the hardware. It did dump all CMD12 requests, though, which
is not correct. This leads to a lock up of the driver because the CMD12
request is never completed.

This patch judges whether the h/w state is in a command sequence or not,
and selects CMD12 handling way for R-Car SDHI.

Signed-off-by: Jan Klötzke <jan.kloetzke@technisat.de>
Signed-off-by: Masaharu Hayakawa <masaharu.hayakawa.ry@renesas.com>
Signed-off-by: Kouei Abe <kouei.abe.cp@renesas.com>
Signed-off-by: Hien Dang <hien.dang.eb@renesas.com>