dma: rcar-audmapp: Change busy check refer bits
This patch will cancel the busy check by the SRS / DRS bit and
correct it to a busy check by the DE bit only.

The SRS / DRS bit should not be changed/cleared at the end control
of audio-dma-pp transfer. This is because the transfer operation
is still in progress when a dma stop request is issued. This patch
is necessary for continued operation when SRS / DRS bit setting is
not cleared.

Signed-off-by: Hiroyuki Yokoyama <hiroyuki.yokoyama.vx@renesas.com>
Signed-off-by: Hien Dang <hien.dang.eb@renesas.com>